requirements.txt
amqp==5.0.6
asgiref==3.3.4
billiard==3.6.4.0
celery==5.1.2
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
Django==3.2
django-celery-beat==2.2.1
django-modeltranslation==0.17.3
django-redis==5.0.0
django-timezone-field==4.2.1
django-widget-tweaks==1.4.8
gunicorn==20.1.0
idna==2.10
kavenegar==1.1.2
kombu==5.1.0
persian-tools==0.0.9
Pillow==8.2.0
prompt-toolkit==3.0.19
psycopg2-binary==2.9.1
python-crontab==2.5.1
python-dateutil==2.8.2
python-decouple==3.4
pytz==2021.1
redis==3.5.3
requests==2.25.1
six==1.16.0
sqlparse==0.4.1
urllib3==1.26.4
vine==5.0.0
wcwidth==0.2.5
